Fix NullEntity Signature and test Uuid generation
NullEntity signature should match other entities, requiring some kind of key (for NullEntity any string will do) for use in generating Uuids when no id is given to the constructor. The pattern is Entity(required1, … , requiredN, uuid=‘’).

Then we add a test that when Uuid is not passed, the entity self-generates a new one, by asserting that an entity created without a passed Uuid matches itself (because it has generated a Uuid that it can match to itself.)